+++ backend/includes/files.h
@@ -11,10 +11,9 @@ typedef struct F {
 } file;
 int open_file(file *f, const char *r_path, int flag);
 int close_file(file *f);
-unsigned long long get_biggest_index(const char *path);
+uint64_t get_biggest_index(const char *path);
 file open_img_at(const char *id, const char *path);
-unsigned long long *open_files_to_arr(const char *path, file *const out,
-                                      uint64_t *files_arr,
-                                      size_t files_arr_size);
+uint64_t *open_files_to_arr(const char *path, file *const out,
+                            uint64_t *files_arr, size_t files_arr_size);
 int attempt_create_dir(const char *path, mode_t mode);
 #endif // FILES_H
